Make HpkeS interface more misuse-resistant
This removes the ability to pass a key pair into the function that
creates the HpkeS instance.  As we are using the basic mode, the
rust-hpke crate doesn't accept a new key pair as that might lead to key
reuse (and massive insecurity).  Instead it generates a fresh one for
every sender instance.  Do the same for NSS.

The consequence is that we can't observe (and log) the details of
the sender key pair when using rust-hpke.  That's OK.  We were only
pretending in the past.  The key pair that was logged was only being
thrown away.

Closes #18.