Move configuration binding to HPKE info
This takes the AAD parameter and includes it in the info string passed
to the export function of HPKE.  This ensures that the request AND the
response are bound to the configuration.

As noted, we don't really have a good reason to do this, but it is safer
if this is how it is bound.  Presently, we don't bind the configuration
to responses.  This leaves us relying the bindings in HPKE, which
already cover the KEM, KDF, and AEAD identifiers.  It doesn't cover the
key identifier, but that is implicitly covered by the choice of key.

That implicit binding isn't ideal, so this approach is - on balance -
safer than what we previously had.