Add support for NSS within application-services
The Mozila application-services library uses this ohttp crate on iOS and uses a
statically linked version of NSS. Update the build.rs to target the same static
NSS build. Note that the NSS_DIR variable in that case is a reorganized version
of the NSS dist directory. Also add an `external-sqlite` crate feature since
application-services provides its own version.