Allow using statically linked softokn for NSS debug builds
The 'pk11wrap' library in NSS uses dlopen to load softokn/freebl libraries at
runtime. As a result we do not need them on the static library list when
building with NSS. Since it may be useful to use static versions of
softokn/freebl, add the option to do so with 'pk11wrap_static'. In this new
mode, we must also name the hardware-specific acceleration libraries.

The default remains to use the dlopen version.