Add unsafe-print-secrets feature flag
This flag is not enabled by default, but it can be used to enable
printing of secrets.  That is, the `PrivateKey` and `SymKey`
implementations will have `Debug` implementations that print the value
of the key if this flag is enabled.

By default, the `Debug` implementations will now just print "Opaque
PrivateKey" or "Opaque SymKey".

Closes #78.